Rename pkpass module to walletpass
I would rather avoid splitting logic about what fields are included in
different mobile wallet passes across two modules, so This Is The Way.
